Adjusted mobile menu logic and top bar content
