+++ index.html
@@ -17,7 +17,7 @@
   <body>
     <div id="app" class="app">
       <div class="top-bar">
-        <p>TOP BAR</p>
+        <p>Nicolas Caballero&reg; - <span id="year-span">2025</span></p>
         <button id="hamburger-button-panel-songs" class="hamburger-button hamburger-button-panel-songs">
           <span class="hamburger-button-line-1"></span>
           <span class="hamburger-button-line-2"></span>
@@ -29,7 +29,7 @@
           <span class="hamburger-button-line-3"></span>
         </button>
       </div>
-      <div class="panel-songs">
+      <div class="panel-songs" data-active="false">
         <div class="song-list-container">
           <h3>ALL SONGS</h3>
           <span class="divider"></span>
@@ -50,7 +50,7 @@
           Upload a Song
         </button>
       </div>
-      <div class="panel-data">
+      <div class="panel-data" data-active="false">
         <div class="properties-container">
           <div class="scenes-container">
             <div class="scenes-dropdown-container">
@@ -86,26 +86,31 @@
                 <input type="color" name="transition-color" id="transition-color-input" required />
                 <label for="background-color-input">Background Color: </label>
                 <input type="color" name="background-color" id="background-color-input" required />
-                <button id="custom-theme-form-button" type="submit" data-add-text="Add Custom Theme" data-update-text="Update Custom Theme">Add Custom Theme</button>
+                <button id="custom-theme-form-button" type="submit" data-add-text="Add Custom Theme" data-update-text="Update Custom Theme">
+                  Add Custom Theme
+                </button>
               </form>
             </div>
             <button id="custom-themes-add-button" data-open="false" data-first-panel-text="Show all Themes" data-second-panel-text="Add a Theme">
               Add a Theme
             </button>
-            <button id="custom-themes-update-button" data-open="false" data-first-panel-text="Show all Themes" data-second-panel-text="Update Selected Theme">
+            <button
+              id="custom-themes-update-button"
+              data-open="false"
+              data-first-panel-text="Show all Themes"
+              data-second-panel-text="Update Selected Theme"
+            >
               Update Selected Theme
             </button>
-            <button id="custom-themes-delete-button">
-              Delete Selected Theme
-            </button>
+            <button id="custom-themes-delete-button">Delete Selected Theme</button>
           </div>
           <span class="divider"></span>
           <label for="enable-rotation-checkbox">Enable camera rotation?</label>
-          <input type="checkbox" name="enable-rotation" id="enable-rotation-checkbox" data-enabled="true"/>
+          <input type="checkbox" name="enable-rotation" id="enable-rotation-checkbox" data-enabled="true" />
           <label for="enable-pan-checkbox">Enable camera pan?</label>
-          <input type="checkbox" name="enable-pan" id="enable-pan-checkbox" data-enabled="true"/>
+          <input type="checkbox" name="enable-pan" id="enable-pan-checkbox" data-enabled="true" />
           <label for="enable-zoom-checkbox">Enable camera zoom?</label>
-          <input type="checkbox" name="enable-zoom" id="enable-zoom-checkbox" data-enabled="true"/>
+          <input type="checkbox" name="enable-zoom" id="enable-zoom-checkbox" data-enabled="true" />
         </div>
       </div>
       <div class="canvas-container">
@@ -141,7 +146,7 @@
         </div>
       </div>
     </div>
-    <!-- <script type="module" src="/src/style.ts"></script> -->
+    <script type="module" src="/src/uiScripts.ts"></script>
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
