+++ src/styles.scss
@@ -82,13 +82,17 @@ canvas {
 
   .panel-songs,
   .panel-data {
+    padding-top: 70px;
+    
     @media screen and (max-width: 768px) {
       position: absolute;
-      top: 0;
+      bottom: 0;
       width: 100%;
-      height: 100%;
+      height: 95%;
       transition: transform 0.3s ease-in-out;
       z-index: 1;
+      padding-top: 50px;
+      
 
       &.active {
         transform: translateX(0%);
@@ -144,6 +148,8 @@ canvas {
     text-align: center;
     width: 100%;
     flex-basis: 100%;
+    font-size: 14px;
+    color: #cdcdcd;
   }
 }
 
@@ -442,7 +448,6 @@ form {
 .properties-container {
   display: flex;
   flex-direction: column;
-  margin: 15px 0;
 }
 
 .panel-songs,
